requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.7.0
boto3==1.35.84
botocore==1.35.84
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
distro==1.9.0
ecdsa==0.19.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.7
langchain-core==0.3.35
langchain-openai==0.3.6
langsmith==0.3.8
MarkupSafe==3.0.2
openai==1.63.0
orjson==3.10.15
packaging==24.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.4
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
wheel==0.44.0
zstandard==0.23.0
